# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ecommerce API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 24 }
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, price_asc, price_desc, popular, rating] }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: inStock
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products, admin]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products, admin]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products, admin]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /products/{id}/related:
    get:
      operationId: listRelatedProducts
      tags: [products]
      summary: List related products
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # ─── Categories ───────────────────────────────────────────────────────────
...
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── Wishlist ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, images, categoryId, inStock, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        categoryName: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        variants: { type: array, items: { $ref: "#/components/schemas/ProductVariant" } }
        inStock: { type: boolean }
        stockCount: { type: ["integer", "null"] }
        isFeatured: { type: boolean }
        averageRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        createdAt: { type: string }
    ProductVariant:
      type: object
      required: [id, name, value]
      properties:
        id: { type: integer }
        name: { type: string }
        value: { type: string }
        priceModifier: { type: ["number", "null"] }
        stockCount: { type: ["integer", "null"] }
    ProductInput:
      type: object
      required: [name, price, categoryId, images]
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: number }
        compareAtPrice: { type: number }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        tags: { type: array, items: { type: string } }
        inStock: { type: boolean }
        stockCount: { type: integer }
        isFeatured: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        compareAtPrice: { type: number }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        tags: { type: array, items: { type: string } }
        inStock: { type: boolean }
        stockCount: { type: integer }
        isFeatured: { type: boolean }
    ProductListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Product" } }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Category ──
...
    Review:
      type: object
      required: [id, productId, rating, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        userId: { type: ["integer", "null"] }
        authorName: { type: string }
        rating: { type: integer }
        title: { type: ["string", "null"] }
        body: { type: ["string", "null"] }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [rating, authorName]
      properties:
        userId: { type: integer }
        authorName: { type: string }
        rating: { type: integer }
        title: { type: string }
        body: { type: string }

    # ── Wishlist ──
...