# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductVariant } from './productVariant';

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  images: string[];
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  tags?: string[];
  variants?: ProductVariant[];
  inStock: boolean;
  /** @nullable */
  stockCount?: number | null;
  isFeatured?: boolean;
  /** @nullable */
  averageRating?: number | null;
  reviewCount?: number;
  createdAt: string;
}