# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug?: string;
  description?: string;
  price: number;
  compareAtPrice?: number;
  images: string[];
  categoryId: number;
  tags?: string[];
  inStock?: boolean;
  stockCount?: number;
  isFeatured?: boolean;
}